+++ src/main/test/java/com/example/demo/dao/impl/TransferStorageTest.java
@@ -10,6 +10,7 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -39,7 +40,7 @@ public class TransferStorageTest extends TestCase {
         transfer.setIdFromBill(2);
         transfer.setIdToUser(15);
         transfer.setIdToBill(3);
-        transfer.setSumTransaction(100);
+        transfer.setSumTransaction(BigDecimal.valueOf(100));
         transfer.setTimeDateTransaction(LocalDateTime.now());
 
         List<Transfer> transferList = new ArrayList<>();
@@ -58,7 +59,7 @@ public class TransferStorageTest extends TestCase {
         firstTransfer.setIdFromBill(2);
         firstTransfer.setIdToUser(15);
         firstTransfer.setIdToBill(3);
-        firstTransfer.setSumTransaction(100);
+        firstTransfer.setSumTransaction(BigDecimal.valueOf(100));
         firstTransfer.setTimeDateTransaction(LocalDateTime.now());
 
         Transfer secondTransfer = new Transfer();
@@ -67,7 +68,7 @@ public class TransferStorageTest extends TestCase {
         secondTransfer.setIdFromBill(3);
         secondTransfer.setIdToUser(5);
         secondTransfer.setIdToBill(1);
-        secondTransfer.setSumTransaction(500);
+        secondTransfer.setSumTransaction(BigDecimal.valueOf(500));
         secondTransfer.setTimeDateTransaction(LocalDateTime.now());
 
         Transfer thirdTransfer = new Transfer();
@@ -76,7 +77,7 @@ public class TransferStorageTest extends TestCase {
         thirdTransfer.setIdFromBill(1);
         thirdTransfer.setIdToUser(2);
         thirdTransfer.setIdToBill(3);
-        thirdTransfer.setSumTransaction(400);
+        thirdTransfer.setSumTransaction(BigDecimal.valueOf(400));
         thirdTransfer.setTimeDateTransaction(LocalDateTime.now());
 
         List<Transfer> transferList = new ArrayList<>();
